Rename imports to reflect github organization url
