+++ src/main/java/io/github/filesystemwatcher/FilesystemWatcher.java
@@ -1,4 +1,4 @@
-package pl.mielecmichal.filesystemmonitor;
+package io.github.filesystemwatcher;
 
 import io.vavr.collection.List;
 import io.vavr.control.Try;
@@ -7,14 +7,7 @@ import lombok.Value;
 import lombok.experimental.NonFinal;
 import lombok.extern.slf4j.Slf4j;
 
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
-import java.nio.file.Watchable;
+import java.nio.file.*;
 import java.util.Map;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
@@ -23,13 +16,8 @@ import java.util.concurrent.Future;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-import static java.nio.file.StandardWatchEventKinds.ENTRY_CREATE;
-import static java.nio.file.StandardWatchEventKinds.ENTRY_DELETE;
-import static java.nio.file.StandardWatchEventKinds.ENTRY_MODIFY;
-import static java.nio.file.StandardWatchEventKinds.OVERFLOW;
-import static pl.mielecmichal.filesystemmonitor.FilesystemEventType.CREATED;
-import static pl.mielecmichal.filesystemmonitor.FilesystemEventType.DELETED;
-import static pl.mielecmichal.filesystemmonitor.FilesystemEventType.INITIAL;
+import static io.github.filesystemwatcher.FilesystemEventType.*;
+import static java.nio.file.StandardWatchEventKinds.*;
 
 @Slf4j
 @Builder
