+++ src/test/java/io/github/filesystemwatcher/parameters/PathKind.java
@@ -1,9 +1,9 @@
-package pl.mielecmichal.filesystemmonitor.parameters;
+package io.github.filesystemwatcher.parameters;
 
+import io.github.filesystemwatcher.utilities.FilesystemUtils;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import pl.mielecmichal.filesystemmonitor.utilities.FilesystemUtils;
 
 import java.nio.file.Path;
 import java.util.List;
