+++ src/test/java/io/github/filesystemwatcher/utilities/FilesystemUtils.java
@@ -1,4 +1,4 @@
-package pl.mielecmichal.filesystemmonitor.utilities;
+package io.github.filesystemwatcher.utilities;
 
 import lombok.Value;
